survey: set the algorithm to get the assigned day
fixes #31

Add a dependency to `moment-business-days` as it has utility functions
to get the holidays, and as long as we use `moment`, we might as well
also use the utility modules that accompany it.

Configure the holidays for the fall 2025 in Quebec.

Set the `_previousDay` in the `home` section's `onSectionEntry` function.

Copy the code for the `serverFieldUpdate` to set the `_assignedDay` when
the `_previousDay` is set from the `od_nationale_2024`. This is a
probabilistic algorithm that checks the probability of being asked the
trips for 1 to 4 days prior to the survey date, depending on the number
of completed interview for each days of the week. To avoid too many
memory bias for the same days, there is always a higher probability
multiplicator for the days closest to the previous day.