webpack: Set the initial title to the survey's title
fixes #75

Use in the first language in the configuration's `languages` field as
the default language and the corresponding `title` for the language as
the default. If it is not set, fallback to the environment variable
`DEFAULT_TITLE` (the one used in the demo_survey webpack) and fallback
again to "Evolution", to avoid having "Webpack app" as the title.

Also set the admin's `noindex` value to `true` to avoid admin apps being
indexed by robots.