sections: rename `tripsSelectPerson` to `selectPerson`
fixes #87

It is the name of the section in `Evolution`. It should match in the
survey, so we can use the select person widgets from Evolution directly.

Also make sure the widgets for active person title and switch person
button are all named `activePersonTitle` and `buttonSwitchPerson`, as
these are the names expected in the templates.