questionnaire: make the interview completed upon reaching section
fixes #76

When entering the `completed` section, if the interview is not marked as
completed, set the `_interviewFinished` response flag to `true`. A
server callback will make the interview completed and set the completion
timestamp. This makes sure that all timestamps are generated server
side.

This is a temporary solution until Evolution takes care of the interview
completion. (see #77)