choices: use `customChoices` when labels have placeholders
fixes #65

The widgets whose choices contained {{}} placeholders now use
customChoices.

Update the choices form the nextPlaceCategory to use context and
placeholders and labels from the excel file instead of those from the
`survey` namespace.

The choices are left as is, even though the widgets use the custom ones,
so that when the issue is fixed and nickname and gender are supported,
the custom can just be removed