UI Tests: Add tripsIntro, end and completed sections for no trips tests with 1 person.
Fix: #12
Test the logout and log back with same credentials for 1 person no trips.
I had to change the label for the two buttons in the end section, because they had the same tests and the tests will failed in these cases.
Fix generateRandomAccessCode to generate access code from 0000-0000 to 9999-9999