UI Tests: Add all sections to two-persons-no-trips tests and add anonymous auth (#108)
* UI Tests: Add all sections to two-persons-no-trips tests

* Auth: Add anonymous auth

It was really difficult to pass the captcha with UI testing.
So I'm adding anonymous authentication until we are done with the UI testing and developing the survey.
We will remove the anonymous auth before going into production.

* Update the accessCode conditionals

Update the accessCode conditionals, because it was not working properly.
Remove the tests for tripsIntro, end and completed sections for 2 persons, it's not working right now.
Update the trips intro section to be for 2 or more persons.