tests: add a task to remove UI test participant data from DB
Add prefill data for participants, with access codes starting with
'7357'.

Add a task that removes all participants whose username starts with
'7357-', which will match the participants of the UI tests. This will
recursively delete the interviews associated with those participants.

Add a script to run with `yarn`: `reset:test:ui`, which runs the task to
remove the UI test participants' interviews. This script can be run in
local setups before running the UI tests to make ensure the data is reset
and tests will not have data already. This avoids having to use the
anonymous login.