config: Remove the anonymous login
fixes #110

Update Evolution with latest ui test fixes

Fix UI tests to use a specific access/postal code pair to avoid having
to enter a captcha.

Add the logout/log back in test to the one-person-no-trips test to make
sure the last visited section (completed) is opened again

Add the `addressIsFilled` parameter to the `fillHomeSectionTests` to not
force entry of home data if the home section is already supposed to be
filled.