ui tests: add the `TRANSITION_DOTENV` to the ui test command
fixes #123

Since the tests now need database access to delete the participant after
the test, the UI tests need the information on the database to use.
Unlike commands running `yarn` where the current direction is kept when
running commands from workspaces, the `npx` command run from the
`test:ui` script takes the workspace directory as the root, so we need
to manually set the .env file path to `../.env` when running the
command.