questionnaire: correctly show the new person popup
fixes #36

This popup is shown in the `tripsIntro` section only, if there are more
than one interviewable person and that person was not manually selected
by the participant in the previous section and the question was not
previously answered for this journey

The value of this popup is set to `true` upon entering the `tripsIntro`
section for the first time for a journey (at journey creation). The widget
will decide whether it should be shown or not in the conditional.