server: do not prefill address if the postal code does not match
fixes #70

The postal code entered at login is retrieved from the username (the
last part of the username after the second '-' character). If it does
not match the postal code in the prefilled data, we do not prefill the
address of the participant (in case they made a typo in the access code,
but have a valid postal code, we don't want to prefill a wrong address
for this interview).

If the username does not match the accessCode/postalCode pattern, or if
the postal code is invalid (the login checked that it is valid, so if
it's invalid, it's not this auth method), for example, for interviews
started by phone interviewers, then the address is prefilled from the
access code as usual.

In case of postalCode/accessCode mismatch, 3 fields are added to the
response: `_postalCodeMismatch`, `_userPostalCode` and
`_prefilledPostalCode` which respectively tell about the mismatch, and
store the postal code entered by the user and the postal code that was
in the pre-filled data.