household: fix household size when person manually added
fixes #146

This changes the conditional for the nickname to be displayed if the
person count is 2 or more instead of the household size, which is not
changed when a person is added in the household section.

Also change the completed conditional of the household section to check
to be `false` only if the reported household size is greater than the
number of persons, as this prevented from navigating to the trips
section when a person was manually added.