selectPerson: show persons in their random order instead of sequential
fixes #143

The list of persons from which to select should match the random order
in which they should be interviewed instead of the sequential order.
Otherwise, when the last person is selected, instead of going to the End
section, it may return to the `selectPerson` section with a previous
person selected. That is confusing.