visitedPlaces: add journey to visited places shortcuts
fixes #85

The code was copy-pasted from 2024, where the concept of journey was not
present. Now, the visited places are under a journey object and their
path should include the journey. The helper function getting the visited
places now returns both the place and the journey it is part of for each
possible place.