travelBehavior: ask no trip reason to all workers and students
fixes #82

All students and workers with fixed study/work location should be asked
the no trip reasons.

Fix UI tests for these questions (add the travelBehavior section to the
UI tests)

Also fix a bug where the `travelBehavior` section was considered as
completed even though there were unanswered questions. We add the helper
function to determine if the travel behavior is complete for a person.