segments: activate busLines questions for Saguenay
part of #63

Add the `busRoutesSaguenay.json` file automatically generated as per the
instructions in the README file from a Transition instance.

Add the implementation of the `busLines` and `busLinesWarning` custom
widgets, copied and adapted from 2024.

Activate the widgets and add missing labels in the Excel file and run
`yarn generateSurvey`.