segments: implement transit summary calculation for Saguenay
fixes #63, fixes #67

Add the scenarios for weekday, saturday and sunday for the saguenay
variant, to allow transit summary calculation

Add a server update callback on the modePre, such that if it is transit,
it will request a transit summary calculation to get the most plausible
transit lines for this trip. This will be calculation only if there are
transit scenarios (ie for the Saguenay variant).

Add unit tests for this function.