tripsIntro: fix the behavior of `personDidTripsChangeConfirm`
fixes #38

Fix the path of the `personDidTrips` answer in the
`personDidTripsChangeConfirm` conditional, as well as the
formattedTripsDate label.

Add a `onSectionExit` function to the `tripsIntro` section to act on the
values of the `personDidTripsChangeConfirm` response. Either reset the
value of `personDidTrips` if the persons did trips after all, or delete
the visited places and trips from the journey.

Make the `visitedPlaces` visible if either personDidTrips or
personDidTripsChangeConfirm is `true`.