household: add gender question if sexAssignedAtBirth preferNotToAnswer
fixes #105

If the answer to `sexAssignedAtBirth` is `preferNotToAnswer`, we ask if
the person would rather tell the gender, which can be custom.

The `gender` field being used to display gender specific labels, it the
sexAssignedAtBirth is `male` or `female`, then it is propagated to the
gender so it can be used in the survey.