segments: add the missing mode choices
fixes #239

Since those modes are possible choices, they need to be present. Their
conditional is set to be displayed only for corresponding `modePre`
values. The question should not be displayed for those modes, as there
will only be one choice.

Also add a warning if there is no mode values for a modePre as it causes
broken images in the template, to better detect if the problem comes
back