server: change assigned day is too far in the past
fixes #59

Add a server field update on the `_sections._actions` field, which is
one field that is always updated upon a new interview activation.

If the last interview update was more than 12 hours ago and the assigned
day is 5 or more days in the past and there are no trips declared in the
interview, then the assigned day is updated to a new date, to reduce the
memory bias if the trips are too far in the past.