visitedPlaces: update condition and button visibility for geography
fixes #234

Use a conditional similar to the one of the `visitedPlaceName` widget to
display the map, ie, hide if the alreadyVisitedPlace is `yes`. But
unlike the name, the map is shown when a shortcut is selected, but
without the search button.

If the location is changed manually by the user, it cancels the shortcut
and the name is visible again.