validations: make sure geographies are not in water
fixes #242

Add the `quebecWaterWaysSimplified.json` file with a feature collection
containing one feature representing the water. It was generated in QGIS
by @kaligrafy, starting from an overpass query to get all
`natural=water` around the Saint-Laurent river in Quebec. Then, only the
rivers and largest lakes were kept to simplify the query. Then in QGis,
a dissolve was done, then an negative buffer of 50m and a 50m
simplification. Finally, the coordinates were updated to reduce
precision to 6 decimals to decrease the size of the resulting geojson.

Points can be drawn really close to the shore, but it's acceptable.

Extract the accessible zone validation in a function and let all
geography widget use this validation function.