Excel: Fixes many issues with Excel Part 6 (#326)
* Label: Remove strike from choice label and value.

Fix: #318

* Label: Update label for household_size

Fix: #289

* Label: Use 2024 for household income year.
We don't want to use lastYear, because we need to wait for the tax, before asking a new year.

Fix: #330

* Label: Revert back to the household_size best label.

* Label: Update 'end' section path, and remove 'skip optional' button

Fix: #213
So, the last questions are optional, except if the respondent click on the range input button, in that case the user need to response to this question in the input range (0 to 100).