visitedPlaces: prevent 2 loop activities in a row
Add a function to validate that the previous and next visited places are
not of specific activity categories.

Call this function in the loop activities conditional to avoid 2 such
activities in a row.

Also call this function for the `workUsual` and `schoolUsual`
activities, as the logic is the same.

This will mitigate #394