admin: do not export `null` ids in simpleExport
fixes #426

The query returned 2 `id` fields, one of which is from the left joined
table, which can be `null` and it was the one retrieved by the query.
Ths `id` field from the sections sub-query is renamed to
`id_with_sections` to avoid the clash and the only `id` is the one from
the main view.

Also sort the interviews by their start date in the export file.