admin: add graph to display interviews by attribute
part of #404

This adds the necessary backend requests and queries to retrieve the
started and completed interviews by any supported attribute.

It adds a dependency to the `recharts` package, which is a `d3` React
wrapper for charts. `highcharts` should eventually be replaced with an
open source charting library and `d3` is a well supported widely used
library, and so is `recharts` package, with 8 millions weekly download
tne active development.

It adds a generic React component to display in a bar chart the started
and completed interview by any attribute. The `attributeName` is passed
as props. The backend route is called with this attribute and the series
are thus built. To properly translate the attribute name and categories,
the translation strings are in the `surveyAdmin.yml` file. There should
be one that is the name of the attribute that will be displayed in the
graph's title. And the categories are under a
`<attributeName>Categories` object.

It supports both horizontal or vertical layout with a prop passed to the
component.

Implement this for the `strate` attribute. The update translation file
is as follows:

```
strateCategories:
  null: "Not specified"
```